1.9.8: Fixes an interesting bug, reported by Benjo Fraser and Roman Mauch.
The bug was caused by the following: to adapt to Big Sur's new look, NSSplitView was changed to the NSViewController-aware mode. To do this, SideOutlineViewController was made into a genuine NSViewController. This added it into the responder chain. Now the contextual menu to add a simple article list called the IBAction addArticleList:, which happens to be implemented in SideOutlineViewController with a different signature, which caused the crash.